Move packages out of pkg directory
